# openapi.yaml
openapi: 3.1.0
info:
  title: FinGood API
  description: |
    AI-Powered Financial Intelligence Platform API
    
    ## Overview
    FinGood provides intelligent transaction categorization and financial analytics for small businesses.
    
    ## Features
    - CSV upload and parsing
    - Rule-based and ML-powered categorization
    - Transaction management and analytics
    - User authentication and authorization
    
    ## Authentication
    This API uses JWT Bearer token authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-token>
    ```
    
    ## Rate Limiting
    - 100 requests per minute per user
    - 1000 requests per hour per user
    
    ## Error Handling
    All errors return a consistent format:
    ```json
    {
      "detail": "Error message",
      "error_code": "ERROR_CODE",
      "timestamp": "2024-01-15T10:30:00Z"
    }
    ```
  version: 1.0.0
  contact:
    name: FinGood Support
    email: support@fingood.com
    url: https://fingood.com/support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  x-logo:
    url: https://fingood.com/logo.png
    altText: FinGood Logo
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.fingood.com
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /api/v1/auth/register:
    post:
      summary: Register a new user
      description: Create a new user account
      operationId: registerUser
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
            examples:
              basic:
                summary: Basic registration
                value:
                  email: "user@example.com"
                  password: "securepassword123"
                  full_name: "John Doe"
                  company_name: "Acme Corp"
                  business_type: "Consulting"
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                email_exists:
                  summary: Email already registered
                  value:
                    detail: "Email already registered"
                    error_code: "EMAIL_EXISTS"
                    timestamp: "2024-01-15T10:30:00Z"
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
  /api/v1/auth/login:
    post:
      summary: Login user
      description: Authenticate user and return access token
      operationId: loginUser
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                  format: email
                  description: User's email address
                password:
                  type: string
                  description: User's password
            examples:
              login:
                summary: User login
                value:
                  username: "user@example.com"
                  password: "securepassword123"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalid_credentials:
                  summary: Invalid email or password
                  value:
                    detail: "Incorrect email or password"
                    error_code: "INVALID_CREDENTIALS"
                    timestamp: "2024-01-15T10:30:00Z"
  /api/v1/auth/me:
    get:
      summary: Get current user
      description: Retrieve current user's profile information
      operationId: getCurrentUser
      tags:
        - Authentication
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/categories:
    get:
      summary: List categories
      description: Retrieve user's categories
      operationId: listCategories
      tags:
        - Categories
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Categories retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategoryResponse'

    post:
      summary: Create category
      description: Create a new category
      operationId: createCategory
      tags:
        - Categories
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryCreate'
      responses:
        '201':
          description: Category created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
  /api/v1/categories/rules:
    get:
      summary: List categorization rules
      description: Retrieve user's categorization rules
      operationId: listCategorizationRules
      tags:
        - Categories
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Rules retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RuleResponse'

    post:
      summary: Create categorization rule
      description: Create a new categorization rule
      operationId: createCategorizationRule
      tags:
        - Categories
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuleCreate'
      responses:
        '201':
          description: Rule created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
    UserCreate:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User's email address
          example: "user@example.com"
        password:
          type: string
          minLength: 8
          description: User's password
          example: "securepassword123"
        full_name:
          type: string
          description: User's full name
          example: "John Doe"
        company_name:
          type: string
          description: Company name
          example: "Acme Corp"
        business_type:
          type: string
          description: Type of business
          example: "Consulting"
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        email:
          type: string
          format: email
          example: "user@example.com"
        full_name:
          type: string
          example: "John Doe"
        company_name:
          type: string
          example: "Acme Corp"
        business_type:
          type: string
          example: "Consulting"
        is_active:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
          description: JWT access token
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        token_type:
          type: string
          example: "bearer"
...
    CategoryCreate:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: Category name
          example: "Office Supplies"
        parent_category:
          type: string
          description: Parent category name
          example: "Business Expenses"
        color:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
          description: Hex color code
          example: "#3b82f6"
    CategoryResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Office Supplies"
        parent_category:
          type: string
          example: "Business Expenses"
        color:
          type: string
          example: "#3b82f6"
        is_default:
          type: boolean
          example: false
        created_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
    RuleCreate:
      type: object
      required:
        - pattern
        - pattern_type
        - category
      properties:
        pattern:
          type: string
          description: Pattern to match (keyword, regex, or vendor name)
          example: "Staples"
        pattern_type:
          type: string
          enum: [keyword, regex, vendor]
          description: Type of pattern matching
          example: "vendor"
        category:
          type: string
          description: Category to assign when pattern matches
          example: "Office Supplies"
        subcategory:
          type: string
          description: Subcategory to assign when pattern matches
          example: "Supplies"
        priority:
          type: integer
          minimum: 1
          maximum: 10
          default: 1
          description: Rule priority (higher numbers applied first)
          example: 1
    RuleResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        pattern:
          type: string
          example: "Staples"
        pattern_type:
          type: string
          enum: [keyword, regex, vendor]
          example: "vendor"
        category:
          type: string
          example: "Office Supplies"
        subcategory:
          type: string
          example: "Supplies"
        priority:
          type: integer
          example: 1
        is_active:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
          description: Human-readable error message
          example: "Transaction not found"
        error_code:
          type: string
          description: Machine-readable error code
          example: "TRANSACTION_NOT_FOUND"
        timestamp:
          type: string
          format: date-time
          description: When the error occurred
          example: "2024-01-15T10:30:00Z"
    ValidationError:
      type: object
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items:
                  oneOf:
                    - type: string
                    - type: integer
                description: Location of the error
              msg:
                type: string
                description: Error message
              type:
                type: string
                description: Error type
          example:
            - loc: ["body", "email"]
              msg: "field required"
              type: "value_error.missing"
...